requirements.txt
langchain>=0.1.5
langchain-community>=0.0.17
langchain-openai>=0.0.5
pyTelegramBotAPI>=4.15.4
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
unstructured[all-docs]>=0.11.2
